modify: rabbitmq配置, modify: redisDocker配置, del: utils的多余代码
