+++ rabbitmq/receiver/mysql_config.json
@@ -1,0 +1,7 @@
+{
+    "username": "root",
+    "password": "12345678",
+    "ip": "mysql_redisshop",
+    "port": "3306",
+    "database": "shop"
+}
