+++ utils/config.go
@@ -1,0 +1,9 @@
+package utils
+
+import (
+	"redisplay/logconf"
+
+	"github.com/sirupsen/logrus"
+)
+
+var logger = logconf.BaseLogger.WithFields(logrus.Fields{"component": "utils"})
